improved hardware
The hardware has been improved and is now more reliable. It does require the phone's microphone to be shorted. This reduces the entire phones impedance to approx 50 Ohm. Allowing for a more reliable dial-pulse detection and a much louder volume from the phones speaker.

Also some easter eggs have been added, dial 002, 003, 008.
The firmware also has been changed to add this new functionality. And the pulse detection is no longer done digitally (which relies on the thresshold of the IO-pin) but analog, so we can determine our own trigger level and create a very large hysteresis swing to be less susceptable to noise.