Menu structure for Volume and MusicFolder configuration
In order to set the volume of the TeleJukebox and to configure which folder with MP3 files (playlist) should be used, a menu has been added. This menu is a classical spoken menu, where the user needs to make it's choice by dialing a number. In order to make this possible, the layout of the SD-card needed to be changed to be able to address all possible folders. This has reduced in some major code changes. To incoporate the state machine for the menu structure.
A timeout has been placed over the number sequence, this way the user cannot accidentally trigger an important settingsmenu or a harmless easter egg. Number sequences need to be dialed with a max delay of 5 seconds per digit, which is very acceptable in practice and prevent a lot of problems.
Also a mesage for dialing 112, 911 or 09008844 has been added.