added english laguage speech files and related emergency number support
Since everything in this project is documented in english, english speech files should be available too.
Also as a fan of "The IT crowd" i could not resists adding the emergency number 0118 999 881 999 119 725 3 to the list of possible emergency numbers. I know it's useless, consider it an easter egg that does nothing special. For those who are not familiar with this number, this may be of help: https://www.youtube.com/watch?v=ab8GtuPdrUQ