oops, wrong names regarding folder 90
the folder names should not contain any text, folder 90 should not be called folder 90 - Nederlands (which is was) in that situation the MP3 player cannot find the folder named 90. Appearently  I did not realize that the naming restrictions are different for folders then for MP3 files. Sorry for any problems this may have caused.