Partslist (in dutch)
This is a partslist describing the costs (in Dutch and made using LibreOffice Calc) of the TeleJukebox.
It contains a list of parts required for assembly of the TeleJukebox and their individual costs. This list was created by one of the builders (not me) but can be of use to others and is therefore hosted here. This list will expire (as prices vary and links will die) and will not be actively supported.